requirements.txt
django>=6.0,<6.1
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=24.0,<25.0
django-cors-headers>=4.4,<5.0
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.2,<3.0
celery[redis]>=5.4,<6.0
redis>=5.0,<6.0
gunicorn>=22.0,<23.0
python-decouple>=3.8,<4.0
whitenoise>=6.7,<7.0
anthropic>=0.40,<1.0
boto3>=1.35,<2.0
weasyprint>=62,<63
plaid-python>=26.0,<27.0
stripe>=10.0,<11.0
openpyxl>=3.1,<4.0
pypdfium2>=4.30,<5.0
rapidocr-onnxruntime>=1.2,<2.0
# ML for the on-device bank-transaction categorizer (replaces Ollama
# calls once we have enough labelled data per entity).
scikit-learn>=1.5,<2.0
joblib>=1.4,<2.0
